Basic game config sorted
The implementation of the config-option sprite (specifically the
button click handling) is super cursed. But it works, and it's not the
point of this project so it'll do.

really we should redo the quip button sprite, as well as provide some
kind of structure for compound sprites, as manually handling updating
sprite children is a pain.